src/requirements.txt
Django>=2.0
Pillow
djangorestframework
djangorestframework-jwt
django-taggit
nltk